Implemented Tiny Mersenne Twister; Reworked random classes

+++ OcularCore/include/Math/Random/ARandom.hpp
@@ -1,41 +1,43 @@
 /**
-* Copyright 2014-2015 Steven T Sell (ssell@ocularinteractive.com)
-*
-* Licensed under the Apache License, Version 2.0 (the "License");
-* you may not use this file except in compliance with the License.
-* You may obtain a copy of the License at
-*
-*    http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-*/
+ * Copyright 2014-2015 Steven T Sell (ssell@ocularinteractive.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
 #pragma once
 #ifndef __H__OCULAR_MATH_ARANDOM__H__
 #define __H__OCULAR_MATH_ARANDOM__H__
 
+#include <cstdint>
+
 //------------------------------------------------------------------------------------------
 
 /**
-* \addtogroup Ocular
-* @{
-*/
+ * \addtogroup Ocular
+ * @{
+ */
 namespace Ocular
 {
     /**
-    * \addtogroup Math
-    * @{
-    */
+     * \addtogroup Math
+     * @{
+     */
     namespace Math
     {
         /**
-        * \addtogroup Random
-        * @{
-        */
+         * \addtogroup Random
+         * @{
+         */
         namespace Random
         {
             /**
@@ -55,37 +57,41 @@ namespace Ocular
 
                 /**
                  * Seeds the PRNG with the specified seed value.
+                 * \param[in] seed
                  */
-                virtual void seed(long long seed);
+                virtual void seed(int64_t seed);
 
                 /**
                  * Retrieves the next pseudo-random number (unbounded).
                  */
-                virtual unsigned next() = 0;
+                virtual uint32_t next() = 0;
 
                 /**
                  * Retrieves the next pseudo-random number and fits it inside of the specified bounds (this is not a clamp)
+                 * 
+                 * \param[in] min Minimum value for the random number
+                 * \param[in] max Maximum value for the random number
                  */
-                virtual unsigned next(unsigned min, unsigned max);
+                virtual uint32_t next(uint32_t min, uint32_t max);
 
             protected:
 
-                long long m_Seed;
+                int64_t m_Seed;
 
             private:
             };
         }
         /**
-        * @} End of Doxygen Groups
-        */
+         * @} End of Doxygen Groups
+         */
     }
     /**
-    * @} End of Doxygen Groups
-    */
+     * @} End of Doxygen Groups
+     */
 }
 /**
-* @} End of Doxygen Groups
-*/
+ * @} End of Doxygen Groups
+ */
 
 //------------------------------------------------------------------------------------------
